+++ classes/WpMatomo.php
@@ -175,7 +175,9 @@ class WpMatomo {
 			}
 		}
 		$tracking_code = new TrackingCode( $this->settings );
-		if ( $this->settings->is_tracking_enabled() && ! $tracking_code->is_hidden_user() ) {
+		if ( $this->settings->is_tracking_enabled()
+		     && $this->settings->get_global_option('track_ecommerce')
+		     && ! $tracking_code->is_hidden_user() ) {
 			$tracker = new AjaxTracker( $this->settings );
 
 			$woocommerce = new Woocommerce( $tracker );
@@ -186,6 +188,8 @@ class WpMatomo {
 
 			$member_press = new MemberPress( $tracker );
 			$member_press->register_hooks();
+
+			do_action( 'matomo_ecommerce_init', $tracker );
 		}
 
 	}